LayoutTests/fast/frames/srcdoc/reloading-a-srcdoc-document-loads-it-again.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

window.iframeLoadCount = 0;

window.onload = function() {
    frames[0].location.reload();
}
</script>
<iframe srcdoc="
  <script>
  ++top.iframeLoadCount;
  alert('Frame has loaded ' + top.iframeLoadCount + ' time(s).');
  if (top.iframeLoadCount == 2) {
    alert('PASS');
    if (window.layoutTestController)
        layoutTestController.notifyDone();
  } else if (top.iframeLoadCount > 2)
    alert('FAIL!!');
  </script>
"></iframe>
